+++ Changelog.md
@@ -1,5 +1,11 @@
 # TORG Eternity Changelog
 
+## NEXT
+
+### Improvements
+
+- Only offer a Destiny Card for a **Defend Approved Action** if the target is doing an "Active Defense" this round.
+
 ## 13.8.2 - Fixes for Perks, Cosms and Secrets
 
 ### Improvements